requirements.txt
huggingface-hub
numpy
pandas
streamlit
tokenizers
torch
transformers
faiss-cpu
sentence_transformers
scipy

